Allow omitting elements during XRC export and import
This is mostly useful for the XRC export to omit elements that are not
available in XRC but was done symmetric for the other direction too.
Since returning nullptr is already an error condition, an empty element
name causes the omitting of the element. This solution ensures plugin
ABI and API compatibility, using an empty name would have created an
invalid output anyway.

+++ src/codegen/xrccg.cpp
@@ -25,6 +25,8 @@
 
 #include "xrccg.h"
 
+#include <cstring>
+
 #include <common/xmlutils.h>
 
 #include "codegen/codewriter.h"
@@ -94,6 +96,13 @@ tinyxml2::XMLElement* XrcCodeGenerator::GetElement(PObjectBase object, tinyxml2:
         }
 
         // Don't return an element for nonvisual objects
+        m_xrc.DeleteNode(objectXrc);
+
+        return nullptr;
+    }
+
+    if (std::strlen(objectXrc->Name()) == 0) {
+        // Special case empty element name, this element shall not be exported to XRC
         m_xrc.DeleteNode(objectXrc);
 
         return nullptr;
